Move Application IT into single class
