+++ integration-tests/src/test/groovy/com/okta/sdk/tests/it/util/ITSupport.groovy
@@ -77,4 +77,8 @@ abstract class ITSupport implements ClientProvider {
 
         return user
     }
+
+    void slowItDown() {
+        Thread.sleep(200)
+    }
 }
